# Fix triggering of Worker Wrapper optimization on Stack