--- a/unison-runtime/unison-runtime.cabal
+++ b/unison-runtime/unison-runtime.cabal
@@ -224,7 +224,6 @@ test-suite runtime-tests
     , mtl
     , primitive
     , stm
-    , template-haskell
     , text
     , unison-core1
     , unison-hash