@pchiusano 2024-11-27
Just curious - I read the precise exceptions page, but I don't know if I get it. It seems like if the function is strict in all its arguments, it shouldn't matter whether you use `throw` or `error`. Is there a function we're calling which isn't fully strict where the distinction matters? Or is it somehow still relevant even if the function is fully strict?