@ChrisPenner 2024-12-06
To be honest I admit I don't fully understand what's happening here or why, just that it in  theory worker-wrapper _should_ trigger with the existing setup but the Core reflects that it doesn't and that this fixes it.

I could spend more time digging in if we think it's important, but changing from a throw to an error seems like a pretty reasonable alternative and adding the test at least makes sure it won't suddenly regress without us noticing :)